# frontend/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditEventOut": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "prev_hash": {
            "title": "Prev Hash",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          }
        },
        "required": [
          "seq",
          "created_at",
          "actor",
          "action",
          "entity_type",
          "entity_id",
          "payload",
          "hash",
          "prev_hash"
        ],
        "title": "AuditEventOut",
        "type": "object"
      },
      "AuditVerifyOut": {
        "properties": {
          "events_checked": {
            "title": "Events Checked",
            "type": "integer"
          },
          "first_bad_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Bad Seq"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "ok",
          "events_checked",
          "first_bad_seq",
          "reason"
        ],
        "title": "AuditVerifyOut",
        "type": "object"
      },
...
      "ConflictAdminOut": {
        "properties": {
          "a": {
            "$ref": "#/components/schemas/ConflictSideOut"
          },
          "b": {
            "$ref": "#/components/schemas/ConflictSideOut"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "detected_by": {
            "title": "Detected By",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "status": {
            "$ref": "#/components/schemas/ConflictStatus"
          }
        },
        "required": [
          "id",
          "description",
          "detected_by",
          "status",
          "confidence",
          "created_at",
          "owner",
          "resolution_note",
          "a",
          "b"
        ],
        "title": "ConflictAdminOut",
        "type": "object"
      },
...
      "ConflictPatch": {
        "properties": {
          "resolution_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "status": {
            "$ref": "#/components/schemas/ConflictStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "ConflictPatch",
        "type": "object"
      },
...
      "ConflictStatus": {
        "enum": [
          "open",
          "resolved",
          "dismissed"
        ],
        "title": "ConflictStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReferenceData": {
        "properties": {
          "branches": {
            "items": {
              "$ref": "#/components/schemas/BranchOut"
            },
            "title": "Branches",
            "type": "array"
          },
          "departments": {
            "items": {
              "$ref": "#/components/schemas/DepartmentOut"
            },
            "title": "Departments",
            "type": "array"
          },
          "doc_types": {
            "items": {
              "type": "string"
            },
            "title": "Doc Types",
            "type": "array"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "title": "Users",
            "type": "array"
          }
        },
        "required": [
          "branches",
          "departments",
          "users",
          "doc_types"
        ],
        "title": "ReferenceData",
        "type": "object"
      },
...
      "StatsOut": {
        "properties": {
          "abstain_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Abstain Reasons",
            "type": "object"
          },
          "abstained": {
            "title": "Abstained",
            "type": "integer"
          },
          "abstention_rate": {
            "title": "Abstention Rate",
            "type": "number"
          },
          "answered": {
            "title": "Answered",
            "type": "integer"
          },
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyPoint"
            },
            "title": "Daily",
            "type": "array"
          },
          "documents_awaiting_approval": {
            "title": "Documents Awaiting Approval",
            "type": "integer"
          },
          "latency_p50_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency P50 Ms"
          },
          "latency_p95_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency P95 Ms"
          },
          "open_conflicts": {
            "title": "Open Conflicts",
            "type": "integer"
          },
          "open_feedback": {
            "title": "Open Feedback",
            "type": "integer"
          },
          "partial": {
            "title": "Partial",
            "type": "integer"
          },
          "pending_supersessions": {
            "title": "Pending Supersessions",
            "type": "integer"
          },
          "route_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Route Counts",
            "type": "object"
          },
          "stale_documents": {
            "items": {
              "$ref": "#/components/schemas/StaleDoc"
            },
            "title": "Stale Documents",
            "type": "array"
          },
          "top_documents": {
            "items": {
              "$ref": "#/components/schemas/CountItem"
            },
            "title": "Top Documents",
            "type": "array"
          },
          "top_questions": {
            "items": {
              "$ref": "#/components/schemas/CountItem"
            },
            "title": "Top Questions",
            "type": "array"
          },
          "total_questions": {
            "title": "Total Questions",
            "type": "integer"
          },
          "unanswered_questions": {
            "items": {
              "$ref": "#/components/schemas/CountItem"
            },
            "title": "Unanswered Questions",
            "type": "array"
          },
          "window_days": {
            "title": "Window Days",
            "type": "integer"
          }
        },
        "required": [
          "window_days",
          "total_questions",
          "answered",
          "partial",
          "abstained",
          "abstention_rate",
          "route_counts",
          "abstain_reasons",
          "latency_p50_ms",
          "latency_p95_ms",
          "top_questions",
          "unanswered_questions",
          "top_documents",
          "stale_documents",
          "open_conflicts",
          "open_feedback",
          "pending_supersessions",
          "documents_awaiting_approval",
          "daily"
        ],
        "title": "StatsOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Source-backed answers from approved institutional documents. ProtoCite retrieves and displays approved documents; it does not provide diagnoses or patient-specific treatment or dosing recommendations.",
    "title": "ProtoCite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/audit": {
      "get": {
        "operationId": "audit_export_api_v1_admin_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventOut"
                  },
                  "title": "Response Audit Export Api V1 Admin Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Export",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/audit/verify": {
      "get": {
        "operationId": "audit_verify_api_v1_admin_audit_verify_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerifyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Verify",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/reference-data": {
      "get": {
        "operationId": "reference_data_api_v1_admin_reference_data_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceData"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reference Data",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/stats": {
      "get": {
        "operationId": "stats_api_v1_admin_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stats",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/conflicts": {
      "get": {
        "operationId": "list_conflicts_api_v1_conflicts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ConflictStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConflictAdminOut"
                  },
                  "title": "Response List Conflicts Api V1 Conflicts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Conflicts",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/conflicts/{conflict_id}": {
      "patch": {
        "operationId": "update_conflict_api_v1_conflicts__conflict_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "conflict_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conflict Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConflictPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Conflict",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}